requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
gunicorn==20.1.0